feat: enhance release deletion process with user authentication and error handling feat: implement MastersClient for improved release display and management fix: adjust delay timings for smoother user experience in batch saving chore: add Toaster component for notifications and alert dialogs chore: update package.json and pnpm-lock.yaml with new dependencies

+++ app/api/db/deleteReleaseById/route.ts
@@ -12,6 +12,40 @@ export async function DELETE(request: NextRequest) {
   }
 
   const supabase = await createClient()
+  if (!supabase) {
+    return NextResponse.json({ error: 'Failed to create Supabase client' }, { status: 500 })
+  }
+
+  // Get the authenticated user
+  const {
+    data: { user },
+  } = await supabase.auth.getUser()
+  if (!user) {
+    return NextResponse.json(
+      { error: 'User must be authenticated to save a master' },
+      { status: 401 }
+    )
+  }
+
+  // check if owner_id matches the user id
+  const { data: release, error: fetchError } = await supabase
+    .from('releases')
+    .select('owner_id')
+    .eq('id', releaseId)
+    .single()
+
+  if (fetchError) {
+    console.error(fetchError.code + ' ' + fetchError.message)
+    return NextResponse.json(
+      { error: 'Error fetching release: ' + fetchError.message },
+      { status: 500 }
+    )
+  }
+
+  if (release.owner_id !== user.id) {
+    return NextResponse.json({ error: 'User is not the owner of this release' }, { status: 403 })
+  }
+
   const { error } = await supabase.from('releases').delete().eq('id', releaseId)
 
   if (error) {
